+++ packages/observable-client/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Stop-recovery: blocks from previous session getting unpinned
 - `operationLimit`-recovery: properly recover from `operationLimit` responses [#494](https://github.com/polkadot-api/polkadot-api/pull/494)
   Closes [#492](https://github.com/polkadot-api/polkadot-api/issues/492)
+- Recover from stop events when runtime hasn't loaded and finalized block changes
 
 ## 0.2.2 - 2024-05-03